Document everything.

+++ src/board.rs
@@ -6,12 +6,14 @@ use crate::{
 };
 
 pub struct Board {
-    sides:  [Bitboard; Nums::SIDES],
-    pieces: [Bitboard; Nums::PIECES],
+    sides:  [Bitboard; Nums::SIDES as usize],
+    pieces: [Bitboard; Nums::PIECES as usize],
     ml:      Movelist,
 }
 
 impl Board {
+    /// Returns a new Board object initialised with the state of the starting
+    /// position.
     pub fn new() -> Board {
         Board {
             sides: [
@@ -30,6 +32,7 @@ impl Board {
         }
     }
 
+    /// Pretty-prints the current state of the board.
     pub fn pretty_print(&self) {
         for r in (Ranks::RANK1..=Ranks::RANK8).rev() {
             print!("{} | ", rank_to_char(r));
@@ -42,10 +45,12 @@ impl Board {
         println!("    1 2 3 4 5 6 7 8");
     }
 
+    /// Returns the piece on the square given by the rank and file, otherwise
+    /// returns '0'.
     fn char_piece_from_pos(&self, rank: u8, file: u8) -> char {
         let sq_bb = bitboard_from_pos(rank, file);
-        for i in 0..Nums::SIDES {
-            for j in 0..Nums::PIECES {
+        for i in 0..Nums::SIDES as usize {
+            for j in 0..Nums::PIECES as usize {
                 if sq_bb & self.sides[i] & self.pieces[j] != 0 {
                     return PIECE_CHARS[i][j];
                 }
@@ -56,6 +61,7 @@ impl Board {
 }
 
 impl Board {
+    /// Runs perft on the given depth.
     pub fn perft(&mut self, depth: u8) -> u8 {
         if depth == 0 {
             return 1;
@@ -72,12 +78,15 @@ impl Board {
         total
     }
 
+    /// REWORK
     pub fn next_move(&mut self) -> Option<Move> {
         self.ml.next()
     }
 }
 
 impl Board {
+    /// REWORK, but supposed to generate all the legal moves from the current
+    /// state of the board.
     pub fn generate_moves(&mut self) {
         // pawn moves
         {
@@ -92,10 +101,12 @@ impl Board {
     }
 }
 
+/// Returns the character representation of the given rank.
 fn rank_to_char(rank: u8) -> char {
     char::from_u32('A' as u32 + rank as u32).unwrap()
 }
 
+/// Returns a 0-initialised bitboard with the bit in the given position set.
 fn bitboard_from_pos(rank: u8, file: u8) -> Bitboard {
     1u64 << (rank * 8 + file)
 }
